[ETCM-126] move tests for personal_XYZ endpoints from JsonRpcControllerSpec into JsonRpcControllerPersonalSpec
